Refactor/restructure
