+++ autoencoder/autoencoder_util.py
@@ -1,0 +1,111 @@
+import torch
+import torchaudio
+
+def make_batches(signal, hp):
+    """
+    signal: Tensor[frame_length] containing the audio data
+    hp: Hyperparameters
+    returns: Tensor[batch_size x (left_context + source + right_context)].
+             A tensor of batches, where each batch is a contiguous slice of audio data. 
+             May be left or right padded with zeros at its ends
+    """
+    source_start_times = torch.arange(start=0, end=signal.shape[0], step=hp.batch_offset, dtype=int)
+    
+    signal = torch.nn.functional.pad(signal, (hp.left_context, hp.source + hp.right_context),
+                                     "constant", 0)
+    subscript_length = hp.left_context + hp.source + hp.right_context
+
+    # make a [batch x subscript_length] tensor for indexing signal
+    indices = torch.arange(start=0, end=subscript_length, step=1, dtype=int)
+    indices = indices.unsqueeze(0).repeat(source_start_times.shape[0], 1)
+    # make a [batch x subscript_length] tensor for shifting indices
+    offset = source_start_times.unsqueeze(1).repeat(1, subscript_length)
+    # add source_start_times[i] to each indices[i]
+    indices += offset
+
+    return signal[indices]
+
+def make_expected_from_batches(batches, hp):
+    """
+    batches: Tensor[batch_count x left_context + source + right_context]. Audio signals
+             for the autoencoder to reproduce
+    hp: Hyperparameters
+
+    returns: Tensor[batch_count x source]. The center slice of each batch we want
+             the autoencoder to reproduce
+    """
+    # make a [source] tensor for subscripting batches
+    indices = torch.arange(start=0, end=hp.source, step=1, dtype=int) + hp.left_context
+
+    result =  batches[:, indices]
+    return result
+
+def put_inference_inside_batches(batches, inference, hp):
+    """
+    batches: Tensor[batch_count x left_context + source + right_context]. Audio signals
+             for the autoencoder to reproduce
+    inference: Tensor[batch_count x source]. The audio signal the autoencoder actually reproduced
+    hp: Hyperparameters
+
+    returns: Tensor[batch_count x left_context + source + right_context]. A copy of batches,
+             with the source samples replaced with the contents of inference
+    """
+    b_copy = batches.clone()
+    # make a [source] tensor for subscripting b_copy
+    indices = torch.arange(start=0, end=hp.source, step=1, dtype=int) + hp.left_context
+    b_copy[:, indices] = inference
+    return b_copy
+
+
+
+class Hyperparameters:
+    """
+    Bag of hyperparameters that need to be passed around at times
+    """
+    def __init__(self,
+                 sample_rate = 16000,
+                 left_context = 16000,
+                 source = 16000,
+                 right_context = 16000,
+                 batch_offset = 16000,
+                 batch_size = 8,
+                 learning_rate = 1e-4,
+                 weight_decay = 1e-5,
+                 epochs = 10,
+                 validate_every_n_batches = 3600,
+                 demo_every_n_epochs = 1,
+                 n_mels = 128,
+                 hop_length = 2000,
+                 n_fft = 4000):
+
+        self.sample_rate = sample_rate
+        self.left_context = left_context
+        self.source = source
+        self.right_context = right_context
+        self.batch_offset = batch_offset
+        self.batch_size = batch_size
+        self.learning_rate = learning_rate
+        self.weight_decay = weight_decay
+        self.epochs = 10
+        self.validate_every_n_batches = validate_every_n_batches
+        self.demo_every_n_epochs = demo_every_n_epochs
+        self.n_mels = n_mels
+        self.hop_length = hop_length
+        self.n_fft = n_fft
+
+        self.device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
+        if torch.cuda.is_available():
+            print("Using CUDA acceleration!")
+
+    def make_spectrogram_object(self):
+        """
+        Makes a MelSpectrogram module
+        """
+        return torchaudio.transforms.MelSpectrogram(
+            sample_rate=self.sample_rate,
+            n_fft=self.n_fft,
+            win_length=self.n_fft,
+            hop_length=self.hop_length,
+            n_mels=self.n_mels,
+            window_fn=torch.hann_window,
+            normalized=True).to(self.device)
