+++ misc/mido-test.py
