+++ misc/model-structure-scratch-work.txt
